Add "Choose colors based on skill" feature to XpGlobes Plugin. 	-Relocated SkillColor.Java from xptracker plugin to game folder (fits in well here as its where the icons get loaded for orbs.)
