+++ runelite-client/src/main/java/net/runelite/client/plugins/xptracker/XpInfoBox.java
@@ -42,6 +42,7 @@ import lombok.Getter;
 import net.runelite.api.Client;
 import net.runelite.api.Experience;
 import net.runelite.api.Skill;
+import net.runelite.client.game.SkillColor;
 import net.runelite.client.game.SkillIconManager;
 import net.runelite.client.ui.ColorScheme;
 import net.runelite.client.ui.DynamicGridLayout;
